requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyYAML==5.3.1
termcolor==1.1.0
Werkzeug==1.0.1
WTForms==2.3.1
